RAG_Based_Job_Matching_System/requirements.txt
langchain
openai
faiss-cpu
llama-index
nltk
spacy
PyPDF2
python-docx
uuid
requests